requirements.txt
# Core pipeline + ASR
faster-whisper==1.0.2

sentencepiece
sox
soundfile

# Text/metrics/compliance
jiwer
presidio-analyzer
presidio-anonymizer
regex
langdetect
pyyaml
scikit-learn
tabulate

# Streamlit + UI
streamlit==1.33.0
streamlit-extras
qdarkstyle
pandas
openpyxl      # <-- required for df.to_excel

# Databases
sqlite-utils

# PDF export
PyPDF2
reportlab

# Transformers/sentiment
transformers==4.40.2
huggingface-hub
tqdm
